Bug: 112104996

+++ src/com/android/managedprovisioning/ota/OtaController.java
@@ -18,12 +18,13 @@ package com.android.managedprovisioning.ota;
 
 import static android.app.admin.DevicePolicyManager.ACTION_PROVISION_MANAGED_DEVICE;
 import static android.app.admin.DevicePolicyManager.ACTION_PROVISION_MANAGED_PROFILE;
+import static android.app.admin.DevicePolicyManager.ACTION_PROVISION_MANAGED_USER;
+
 import static com.android.internal.util.Preconditions.checkNotNull;
 
 import android.app.admin.DevicePolicyManager;
 import android.content.ComponentName;
 import android.content.Context;
-import android.content.pm.PackageManager;
 import android.content.pm.UserInfo;
 import android.os.UserHandle;
 import android.os.UserManager;
@@ -36,6 +37,7 @@ import com.android.managedprovisioning.task.DeleteNonRequiredAppsTask;
 import com.android.managedprovisioning.task.DisableInstallShortcutListenersTask;
 import com.android.managedprovisioning.task.DisallowAddUserTask;
 import com.android.managedprovisioning.task.InstallExistingPackageTask;
+import com.android.managedprovisioning.task.MigrateSystemAppsSnapshotTask;
 
 /**
  * After a system update, this class resets the cross-profile intent filters and performs any
@@ -72,6 +74,9 @@ public class OtaController {
         if (mContext.getUserId() != UserHandle.USER_SYSTEM) {
             return;
         }
+        // Migrate snapshot files to use user serial number as file name.
+        mTaskExecutor.execute(
+                UserHandle.USER_SYSTEM, new MigrateSystemAppsSnapshotTask(mContext, mTaskExecutor));
 
         // Check for device owner.
         final int deviceOwnerUserId = mDevicePolicyManager.getDeviceOwnerUserId();
@@ -82,6 +87,8 @@ public class OtaController {
         for (UserInfo userInfo : mUserManager.getUsers()) {
             if (userInfo.isManagedProfile()) {
                 addManagedProfileTasks(userInfo.id, mContext);
+            } else if (mDevicePolicyManager.getProfileOwnerAsUser(userInfo.id) != null) {
+                addManagedUserTasks(userInfo.id, mContext);
             } else {
                 // if this user has managed profiles, reset the cross-profile intent filters between
                 // this user and its managed profiles.
@@ -131,6 +138,23 @@ public class OtaController {
                 .build();
         mTaskExecutor.execute(userId,
                 new DisableInstallShortcutListenersTask(context, fakeParams, mTaskExecutor));
+        mTaskExecutor.execute(userId,
+                new DeleteNonRequiredAppsTask(false, context, fakeParams, mTaskExecutor));
+    }
+
+    void addManagedUserTasks(final int userId, Context context) {
+        ComponentName profileOwner = mDevicePolicyManager.getProfileOwnerAsUser(userId);
+        if (profileOwner == null) {
+            // Shouldn't happen.
+            ProvisionLogger.loge("No profile owner on managed user " + userId);
+            return;
+        }
+
+        // Build a set of fake params to be able to run the tasks
+        ProvisioningParams fakeParams = new ProvisioningParams.Builder()
+                .setDeviceAdminComponentName(profileOwner)
+                .setProvisioningAction(ACTION_PROVISION_MANAGED_USER)
+                .build();
         mTaskExecutor.execute(userId,
                 new DeleteNonRequiredAppsTask(false, context, fakeParams, mTaskExecutor));
     }
